Add generator for domain
Added a generator for the domain in the
that was causing an issue with the polynomial
interpolation for the Reed-Solomon code. Now the
verifier can correctly reconstruct the polynomial
and check that it is of low degree.

This proves that it is a valid codeword, and
now we only have to implement the statemachine to
create the correct polynomial.